# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: HITCON 2026 NFC Battle API
  version: 1.0.0
  description: |
    OpenAPI specification for the HITCON 2026 NFC Battle backend.

    `Backend/game-flow.md` is the source of truth for product behavior. This
    file mirrors that flow as a machine-readable API contract.

    User roles:
    - `ATTENDEE`
    - `STAFF`
    - `SPONSOR`
    - `COMMUNITY`

    Authenticated app APIs use the conference JWT:
    `Authorization: Bearer <JWT>`. The backend derives caller identity from the
    JWT subject, which is the user ID provided by the JWT issuer. JWT
    verification uses a shared secret and HMAC.

    Required JWT claims:
    - `sub`: user ID, provided by the JWT issuer.
    - `exp`: expiration time.
    - `iss`: issuer.
    - `aud`: audience.
    - `role`: caller role, used for fast role lookup without querying the DB.

    Physical NFC tag IDs are predictable, so scan requests must include both the
    parsed user ID from the tag URL and the physical ID read from the tag. The
    server must verify that the physical ID belongs to the parsed user ID before
    adding the target user to the scanner's collection.

    Collection records are the long-term permission for full profile access.
    `profile_version` and `collection_version` are integer versions stored on
    the user row. `profile_version` changes when profile fields change.
    `collection_version` changes only when that user's own collection changes.

...
    reopens scoring.

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /users/me:
    get:
      tags: [Users]
      summary: Lazy initialize or get my profile
      description: |
        Creates a default profile when the authenticated JWT subject is new,
        then returns the authenticated user's complete data. If the user has
        not paired an NFC tag yet, `physical_id` is returned as `null`.
      responses:
        '200':
          description: Profile loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyProfileResponse'
              examples:
                profile:
                  $ref: '#/components/examples/GetMyProfileResponse'
                unpaired:
                  $ref: '#/components/examples/GetMyUnpairedProfileResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [Users]
      summary: Update my profile
      description: |
        Partially updates the authenticated user's profile. The user may send
        only the fields they want to change. Updatable fields are
        `display_name`, `emoji_icon`, `bio`, and `pixel_avatar_base64`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMyProfileRequest'
            examples:
              update:
                $ref: '#/components/examples/UpdateMyProfileRequest'
      responses:
        '200':
          description: Profile updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyProfileResponse'
              examples:
                profile:
                  $ref: '#/components/examples/GetMyProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/me/prize:
    get:
      tags: [Users]
      summary: Get my prize result
      description: |
        Returns the authenticated user's prize result after the scoreboard is
        frozen and prize calculation has run. Prize results are read from the
        stored freeze snapshot. This endpoint is rejected while the scoreboard
        is `OPEN` or `FREEZING`.
      responses:
        '200':
          description: Prize result loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyPrizeResponse'
              examples:
                prize:
                  $ref: '#/components/examples/GetMyPrizeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ScoreboardNotFrozenError'
  /users/me/bootstrap:
    get:
      tags: [Users]
      summary: Bootstrap my local app cache
      description: |
        Returns the authenticated user's full profile plus full profiles for all
        users they have collected. This endpoint is intended for app recovery or
        fresh installs, not normal incremental refresh. Collected user profiles
        do not include nested collection lists.
      responses:
        '200':
          description: Bootstrap data loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMyBootstrapResponse'
              examples:
                bootstrap:
                  $ref: '#/components/examples/GetMyBootstrapResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/batch:
    post:
      tags: [Users]
      summary: Batch refresh user profiles
      description: |
        Read-only, idempotent batch lookup for multiple cached user profiles.
        This endpoint uses `POST` only so the client can send a structured
        request body with many user IDs and cached versions.

        For each user, the server applies the same visibility rule as
        `GET /users/{user_id}`: full profile if the caller has collected that
        user, otherwise partial profile. If the client sends both cached
        versions and both match the server's current versions, the server
        returns only `user_id` and `unchanged: true` for that user.

        `profile_version` is the client's cached version of that user's profile
        fields. `collection_version` is the client's cached version of that
        user's collection list. When either cached version is omitted or stale,
        the server returns `unchanged: false` plus the user's profile using the
        same visibility behavior as `GET /users/{user_id}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGetUsersRequest'
            examples:
              batch:
                $ref: '#/components/examples/BatchGetUsersRequest'
      responses:
        '200':
          description: Batch profile data loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGetUsersResponse'
              examples:
                batch:
                  $ref: '#/components/examples/BatchGetUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/UserNotFoundError'
  /users/{user_id}:
    get:
      tags: [Users]
      summary: Get another user's profile
      description: |
        Returns another user's profile. If the caller has collected the queried
        user, the server returns full profile data plus version metadata.
        Otherwise, it only returns `user_id`, `display_name`, and `emoji_icon`,
        allowing the mobile app to show a locked character state.

        The client may send both `profile_version` and `collection_version` for
        a cached full profile. If both versions match the server's current
        versions, the server returns an unchanged marker instead of the profile
        body. If either version is omitted or stale, the server returns the
        profile body using the normal visibility rule.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: Target user ID from the JWT issuer.
          example: kktix_hash_abc123
        - in: query
          name: profile_version
          required: false
          schema:
            type: integer
            minimum: 0
          description: Locally cached profile version for this user.
          example: 2
        - in: query
          name: collection_version
          required: false
          schema:
            type: integer
            minimum: 0
          description: Locally cached collection version for this user.
          example: 5
      responses:
        '200':
          description: Profile loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserProfileResponse'
              examples:
                partial:
                  $ref: '#/components/examples/GetPartialUserProfileResponse'
                full:
                  $ref: '#/components/examples/GetFullUserProfileResponse'
                unchanged:
                  $ref: '#/components/examples/GetUnchangedUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/UserNotFoundError'
  /users/{user_id}/collection:
    get:
      tags: [Users]
      summary: Get another user's collection
      description: |
        Returns the queried user's collection only when the caller has collected
        the queried user. For each user in that collection, the server applies
        the same visibility rule from the caller's point of view: full profile
        if the caller has collected that listed user, otherwise partial profile.

        The client may send `collection_version` for a cached collection. If it
        matches the server's current collection version for the queried user,
        the server returns an unchanged marker instead of the collection body.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          description: Target user ID from the JWT issuer.
          example: kktix_hash_def456
        - in: query
          name: collection_version
          required: false
          schema:
            type: integer
            minimum: 0
          description: Locally cached collection version for this user.
          example: 5
      responses:
        '200':
          description: Collection loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserCollectionResponse'
              examples:
                collection:
                  $ref: '#/components/examples/GetUserCollectionResponse'
                unchanged:
                  $ref: '#/components/examples/GetUnchangedUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/UserNotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Conference JWT sent as `Authorization: Bearer <JWT>`. Verified with a shared secret and HMAC. Required claims: `sub`, `exp`, `iss`, `aud`, and `role`.'
    staffDangerToken:
      type: apiKey
      in: header
      name: STAFF_DANGER_TOKEN
      description: Staff danger token for freeze/resume operations.
  responses:
    BadRequestError:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            code: BAD_REQUEST
            message: Invalid request body or query parameter.
    UnauthorizedError:
      description: Missing or invalid HMAC-signed JWT.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            code: UNAUTHORIZED
            message: Missing or invalid JWT.
...
    UserNotFoundError:
      description: User not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            code: USER_NOT_FOUND
            message: User not found.
    ForbiddenError:
      description: Forbidden.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            code: FORBIDDEN
            message: Forbidden.
...
    ScoreboardNotFrozenError:
      description: Scoreboard is not frozen yet.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            code: SCOREBOARD_NOT_FROZEN
            message: Scoreboard is not frozen yet.
  schemas:
    ErrorResponse:
      type: object
      required: [status, code, message]
      properties:
        status:
          type: string
          enum: [error]
        code:
          type: string
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - STAFF_DANGER_TOKEN_INVALID
            - FORBIDDEN
            - USER_NOT_FOUND
            - PHYSICAL_ID_MISMATCH
            - TAG_ALREADY_PAIRED
            - SCOREBOARD_ALREADY_FROZEN
            - SCOREBOARD_FREEZING
            - SCOREBOARD_NOT_FROZEN
        message:
          type: string
...
    UserRole:
      type: string
      enum: [ATTENDEE, STAFF, SPONSOR, COMMUNITY]
    UserProfilePublicFull:
      type: object
      required:
        - user_id
        - display_name
        - role
        - emoji_icon
        - bio
        - pixel_avatar_base64
        - profile_version
        - collection_version
      properties:
        user_id:
          type: string
          description: User ID provided by the JWT issuer.
          example: kktix_hash_abc123
        display_name:
          type: string
          example: Hacker_Aries
        role:
          $ref: '#/components/schemas/UserRole'
        emoji_icon:
          type: string
          description: Emoji character shown as the user's icon.
          example: 🐱
        bio:
          type: string
          example: I love reverse engineering.
        pixel_avatar_base64:
          type: string
          description: Base64-encoded pixel avatar image.
          example: iVBORw0KGgoAAAANSU...
        profile_version:
          type: integer
          description: Incremented when profile fields change.
          example: 3
        collection_version:
          type: integer
          description: Incremented when this user's own collection changes.
          example: 8
    UserProfileSelf:
      allOf:
        - $ref: '#/components/schemas/UserProfilePublicFull'
        - type: object
          required:
            - physical_id
            - collection
          properties:
            physical_id:
              type: string
              nullable: true
              description: Paired NFC tag physical ID. This is null before the user pairs an NFC tag.
              example: '04:1A:2B:3C:4D:5E:6F'
            collection:
              type: array
              description: User IDs this user has previously scanned.
              items:
                type: string
              example: [kktix_hash_def456]
    UserProfilePartial:
      type: object
      required: [user_id, display_name, emoji_icon]
      properties:
        user_id:
          type: string
          example: kktix_hash_def456
        display_name:
          type: string
          example: Alice
        emoji_icon:
          type: string
          description: Emoji character shown as the user's icon.
          example: 😉
    UserCollectionItem:
      oneOf:
        - $ref: '#/components/schemas/UserProfilePublicFull'
        - $ref: '#/components/schemas/UserProfilePartial'
    UserCollection:
      type: object
      required: [user_id, collection_version, users]
      properties:
        user_id:
          type: string
          description: Owner of this collection.
          example: kktix_hash_def456
        collection_version:
          type: integer
          example: 8
        users:
          type: array
          description: Hydrated users in the collection.
          items:
            $ref: '#/components/schemas/UserCollectionItem'
    GetMyProfileResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/UserProfileSelf'
    GetMyBootstrapResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          required: [me, collected_users]
          properties:
            me:
              $ref: '#/components/schemas/UserProfileSelf'
            collected_users:
              type: array
              description: Full profiles for users the authenticated user has collected.
              items:
                $ref: '#/components/schemas/UserProfilePublicFull'
    UpdateMyProfileRequest:
      type: object
      description: |
        Partial profile update. Only `display_name`, `emoji_icon`, `bio`, and
        `pixel_avatar_base64` are updatable.
      additionalProperties: false
      properties:
        display_name:
          type: string
          example: Aries_The_Great
        emoji_icon:
          type: string
          description: Emoji character shown as the user's icon.
          example: 🚀
        bio:
          type: string
          example: Updated bio.
        pixel_avatar_base64:
          type: string
          example: iVBORw0KGgo...
    GetUserProfileResponse:
      oneOf:
        - $ref: '#/components/schemas/GetUnchangedUserResponse'
        - $ref: '#/components/schemas/GetPartialUserProfileResponse'
        - $ref: '#/components/schemas/GetFullUserProfileResponse'
    CachedUserUnchanged:
      type: object
      required: [user_id, unchanged]
      properties:
        user_id:
          type: string
          example: kktix_hash_def456
        unchanged:
          type: boolean
          enum: [true]
    GetUnchangedUserResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/CachedUserUnchanged'
    GetPartialUserProfileResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/UserProfilePartial'
    GetFullUserProfileResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/UserProfilePublicFull'
    GetUserCollectionResponse:
      oneOf:
        - $ref: '#/components/schemas/GetUserCollectionLoadedResponse'
        - $ref: '#/components/schemas/GetUnchangedUserResponse'
    GetUserCollectionLoadedResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/UserCollection'
    BatchGetUsersRequest:
      type: object
      required: [users]
      additionalProperties: false
      properties:
        users:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [user_id]
            additionalProperties: false
            properties:
              user_id:
                type: string
                example: kktix_hash_def456
              profile_version:
                type: integer
                minimum: 0
                description: |
                  Locally cached profile version for this user, if any. If this
                  is omitted, the server treats the profile cache as stale.
                example: 2
              collection_version:
                type: integer
                minimum: 0
                description: |
                  Locally cached collection version for this user's collection
                  list, if any. If the caller has full access to this user and
                  this is omitted, the server treats the collection cache as
                  stale.
                example: 7
    BatchUserResult:
      oneOf:
        - $ref: '#/components/schemas/CachedUserUnchanged'
        - type: object
          required: [user_id, unchanged, data]
          properties:
            user_id:
              type: string
              example: kktix_hash_def456
            unchanged:
              type: boolean
              enum: [false]
            data:
              description: Profile data using the same visibility behavior as `GET /users/{user_id}`.
              oneOf:
                - $ref: '#/components/schemas/UserProfilePublicFull'
                - $ref: '#/components/schemas/UserProfilePartial'
    BatchGetUsersResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          required: [results]
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/BatchUserResult'
...
    GetMyPrizeResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          required: [scoreboard_frozen, stamp_prize, rank_prize]
          properties:
            scoreboard_frozen:
              type: boolean
              enum: [true]
            stamp_prize:
              type: boolean
              description: Whether the user won the stamp prize.
              example: true
            rank_prize:
              type: boolean
              description: Whether the user won the rank prize.
              example: true
            rank:
              type: integer
              nullable: true
              example: 7
...
  examples:
    GetMyProfileResponse:
      value:
        status: success
        data:
          user_id: kktix_hash_abc123
          display_name: Hacker_Aries
          role: ATTENDEE
          emoji_icon: 🐱
          bio: I love reverse engineering.
          pixel_avatar_base64: iVBORw0KGgoAAAANSU...
          profile_version: 3
          collection_version: 12
          physical_id: '04:1A:2B:3C:4D:5E:6F'
          collection:
            - kktix_hash_def456
    GetMyUnpairedProfileResponse:
      value:
        status: success
        data:
          user_id: kktix_hash_new_user
          display_name: Player_new_user
          role: ATTENDEE
          emoji_icon: 🙂
          bio: ''
          pixel_avatar_base64: ''
          profile_version: 1
          collection_version: 0
          physical_id: null
          collection: []
    GetMyBootstrapResponse:
      value:
        status: success
        data:
          me:
            user_id: kktix_hash_abc123
            display_name: Hacker_Aries
            role: ATTENDEE
            emoji_icon: 🐱
            bio: I love reverse engineering.
            pixel_avatar_base64: iVBORw0KGgoAAAANSU...
            profile_version: 3
            collection_version: 12
            physical_id: '04:1A:2B:3C:4D:5E:6F'
            collection:
              - kktix_hash_def456
          collected_users:
            - user_id: kktix_hash_def456
              display_name: Plant_Lover
              role: COMMUNITY
              emoji_icon: 🌱
              bio: Ask me about community projects.
              pixel_avatar_base64: iVBORw0KGgoAAAANSU...
              profile_version: 2
              collection_version: 5
    UpdateMyProfileRequest:
      value:
        display_name: Aries_The_Great
        emoji_icon: 🚀
        bio: Updated bio.
        pixel_avatar_base64: iVBORw0KGgo...
    BatchGetUsersRequest:
      value:
        users:
          - user_id: kktix_hash_def456
            profile_version: 2
            collection_version: 5
          - user_id: kktix_hash_ctf
            profile_version: 1
    BatchGetUsersResponse:
      value:
        status: success
        data:
          results:
            - user_id: kktix_hash_def456
              unchanged: true
            - user_id: kktix_hash_ctf
              unchanged: false
              data:
                user_id: kktix_hash_ctf
                display_name: CTF_Player
                emoji_icon: 🐱
    GetPartialUserProfileResponse:
      value:
        status: success
        data:
          user_id: kktix_hash_def456
          display_name: Alice
          emoji_icon: 😉
    GetFullUserProfileResponse:
      value:
        status: success
        data:
          user_id: kktix_hash_def456
          display_name: Plant_Lover
          role: COMMUNITY
          emoji_icon: 🌱
          bio: Ask me about community projects.
          pixel_avatar_base64: iVBORw0KGgoAAAANSU...
          profile_version: 2
          collection_version: 5
    GetUnchangedUserResponse:
      value:
        status: success
        data:
          user_id: kktix_hash_def456
          unchanged: true
    GetUserCollectionResponse:
      value:
        status: success
        data:
          user_id: kktix_hash_def456
          collection_version: 5
          users:
            - user_id: kktix_hash_ctf
              display_name: CTF_Player
              emoji_icon: 🐱
            - user_id: kktix_hash_root
              display_name: Root_User
              role: STAFF
              emoji_icon: 💻
              bio: Root shell collector.
              pixel_avatar_base64: iVBORw0KGgoAAAANSU...
              profile_version: 4
              collection_version: 10
...
    GetMyPrizeResponse:
      value:
        status: success
        data:
          scoreboard_frozen: true
          stamp_prize: true
          rank_prize: true
          rank: 7
...